Refactored the code in order to correct errors & warnings from audit tools 'checkstyle, findbugs, cpd ...'

+++ paissadtools-utils/src/main/java/net/paissad/paissadtools/util/ZipUtils.java
@@ -16,6 +16,7 @@ import java.util.zip.ZipFile;
 import java.util.zip.ZipOutputStream;
 
 import org.apache.commons.io.FileUtils;
+import org.apache.commons.io.IOUtils;
 
 /**
  * This class contains mainly methods that help to compress / uncompress files
@@ -23,9 +24,8 @@ import org.apache.commons.io.FileUtils;
  * dependency.
  * 
  * @author paissad
- * @version 0.1
  */
-public class ZipUtils {
+public final class ZipUtils {
 
     private static final int    BUFFER       = 4096;
 
@@ -35,13 +35,16 @@ public class ZipUtils {
 
     private static final String FILE_SEP     = File.separator;
 
+    private ZipUtils() {
+    }
+
     /**
      * Compress a file or directory into a .zip file. If the destination file
      * does already exist, it is overwritten.
      * 
      * @param source - The source file or directory to compress.
      * @param destZipFile - The zip file to create.
-     * @throws IOException
+     * @throws IOException If an error occurs while zipping the file.
      */
     public static void zip(final File source, final File destZipFile) throws IOException {
 
@@ -51,52 +54,45 @@ public class ZipUtils {
         try {
             zos = new ZipOutputStream(new BufferedOutputStream(new FileOutputStream(destZipFile)));
             // out.setMethod(ZipOutputStream.DEFLATED);
-            byte data[] = new byte[BUFFER];
-            int bytesRead;
             String entryName;
 
             if (source.isFile()) { // If the source is a file.
                 entryName = source.getName();
                 bis = new BufferedInputStream(new FileInputStream(source), BUFFER);
-                ZipEntry entry = new ZipEntry(entryName);
+                final ZipEntry entry = new ZipEntry(entryName);
                 entry.setTime(source.lastModified());
                 zos.putNextEntry(entry);
-                while ((bytesRead = bis.read(data, 0, BUFFER)) != -1) {
-                    zos.write(data, 0, bytesRead);
-                }
+                IOUtils.copy(bis, zos);
                 bis.close();
 
             } else { // If the source is a directory.
-                List<File> filesToZip = new ArrayList<File>();
+                final List<File> filesToZip = new ArrayList<File>();
                 getChildFiles(source, filesToZip);
                 if (!filesToZip.isEmpty()) {
-                    for (File aFile : filesToZip) {
+                    for (final File aFile : filesToZip) {
                         // We take the relative path of the file
-                        entryName = aFile.getAbsolutePath().substring(
-                                source.getAbsolutePath().length() + 1);
+                        entryName = aFile.getAbsolutePath().substring(source.getAbsolutePath().length() + 1);
                         // Use the '/' as file separator.
                         entryName = entryName.replace(File.separatorChar, ZIP_SEP_CHAR);
                         if (aFile.isFile()) {
                             bis = new BufferedInputStream(new FileInputStream(aFile), BUFFER);
-                            ZipEntry entry = new ZipEntry(entryName);
+                            final ZipEntry entry = new ZipEntry(entryName);
                             entry.setTime(aFile.lastModified());
                             zos.putNextEntry(entry);
-                            while ((bytesRead = bis.read(data, 0, BUFFER)) != -1) {
-                                zos.write(data, 0, bytesRead);
-                            }
+                            IOUtils.copy(bis, zos);
                             bis.close();
                         } else {
                             // If the directory is empty, we add an '/' at the
                             // end of the entry's name.
                             entryName += "/";
-                            ZipEntry entry = new ZipEntry(entryName);
+                            final ZipEntry entry = new ZipEntry(entryName);
                             entry.setTime(aFile.lastModified());
                             zos.putNextEntry(entry);
                         }
                     }
 
                 } else { // If the directory is empty.
-                    ZipEntry emptyEntry = new ZipEntry("./");
+                    final ZipEntry emptyEntry = new ZipEntry("./");
                     emptyEntry.setTime(source.lastModified());
                     zos.putNextEntry(emptyEntry);
                 }
@@ -121,7 +117,8 @@ public class ZipUtils {
      * @throws IOException If a problem occurs during the uncompressing process.
      * @see #unzip(File, File, boolean)
      */
-    public static void unzip(File sourceZip, File destinationDir) throws IllegalArgumentException, IOException {
+    public static void unzip(final File sourceZip, final File destinationDir) throws IllegalArgumentException,
+            IOException {
         unzip(sourceZip, destinationDir, true);
     }
 
@@ -143,45 +140,45 @@ public class ZipUtils {
      * @throws IOException If a problem occurs during the uncompressing process.
      * @see #unzip(File, File)
      */
-    public static void unzip(final File sourceZip, final File destDir, boolean isDeleteDestDir)
+    public static void unzip(final File sourceZip, final File destDir, final boolean isDeleteDestDir)
             throws IllegalArgumentException, IOException {
 
         BufferedInputStream bis = null;
         BufferedOutputStream bos = null;
+        ZipFile zipFile = null;
 
         try {
             // Remove the destination directory if specified.
             if (isDeleteDestDir) {
                 if (destDir == null) throw new IllegalArgumentException("The destination directory cannot be null");
 
-                if (destDir.exists())
-                    FileUtils.forceDelete(destDir);
-                destDir.mkdirs();
+                if (destDir.exists()) FileUtils.forceDelete(destDir);
+                FileUtils.forceMkdir(destDir);
             } else {
-                if (!destDir.exists()) destDir.mkdirs();
+                if (!destDir.exists()) FileUtils.forceMkdir(destDir);
             }
 
-            ZipFile zipFile = new ZipFile(sourceZip, ZipFile.OPEN_READ);
-            Enumeration<ZipEntry> entries = (Enumeration<ZipEntry>) zipFile.entries();
+            zipFile = new ZipFile(sourceZip, ZipFile.OPEN_READ);
+            final Enumeration<ZipEntry> entries = (Enumeration<ZipEntry>) zipFile.entries();
             int count;
-            byte[] data = new byte[BUFFER];
+            final byte[] data = new byte[BUFFER];
 
             while (entries.hasMoreElements()) {
-                ZipEntry aEntry = entries.nextElement();
-                String entryName = aEntry.getName();
+                final ZipEntry aEntry = entries.nextElement();
+                final String entryName = aEntry.getName();
                 if (aEntry.isDirectory()) {
-                    new File(destDir.getAbsolutePath() + FILE_SEP + entryName).mkdirs();
+                    FileUtils.forceMkdir(new File(destDir.getAbsolutePath() + FILE_SEP + entryName));
                 } else {
                     // Regexp that permits to retrieve the name of the parent
                     // directory of the current entry.
-                    Pattern pattern = Pattern.compile(".*/|.*\\\\");
-                    Matcher matcher = pattern.matcher(entryName);
+                    final Pattern pattern = Pattern.compile(".*/|.*\\\\");
+                    final Matcher matcher = pattern.matcher(entryName);
                     // We retrieve the name of the parent's directory
-                    String parentEntryName = (matcher.find()) ? matcher.group() : "";
-                    new File(destDir.getAbsolutePath() + FILE_SEP + parentEntryName).mkdirs();
+                    final String parentEntryName = (matcher.find()) ? matcher.group() : "";
+                    FileUtils.forceMkdir(new File(destDir.getAbsolutePath() + FILE_SEP + parentEntryName));
                     bis = new BufferedInputStream(zipFile.getInputStream(aEntry), BUFFER);
-                    bos = new BufferedOutputStream(new FileOutputStream(destDir.getAbsolutePath()
-                            + FILE_SEP + entryName), BUFFER);
+                    bos = new BufferedOutputStream(new FileOutputStream(destDir.getAbsolutePath() + FILE_SEP
+                            + entryName), BUFFER);
                     while ((count = bis.read(data, 0, BUFFER)) != -1)
                         bos.write(data, 0, count);
                     bos.flush();
@@ -191,6 +188,7 @@ public class ZipUtils {
 
         } finally {
             CommonUtils.closeAllStreamsQuietly(bis, bos);
+            if (zipFile != null) zipFile.close();
         }
     }
 
@@ -212,7 +210,7 @@ public class ZipUtils {
         }
     }
 
-    private static boolean isEmptyDir(File directory) throws IllegalArgumentException, IllegalStateException {
+    private static boolean isEmptyDir(final File directory) throws IllegalArgumentException, IllegalStateException {
         if (!directory.isDirectory()) throw new IllegalArgumentException("Is not a directory");
         if (!directory.canRead()) throw new IllegalStateException("Cannot read the directory.");
         if (directory.list() == null) throw new IllegalStateException("Error while reading directory");
